fixed mediumint was mapped on smallint, mapped now to the new type mediumint

+++ generator/lib/reverse/mysql/MysqlSchemaParser.php
@@ -32,7 +32,7 @@ class MysqlSchemaParser extends BaseSchemaParser
     private static $mysqlTypeMap = array(
         'tinyint' => PropelTypes::TINYINT,
         'smallint' => PropelTypes::SMALLINT,
-        'mediumint' => PropelTypes::SMALLINT,
+        'mediumint' => PropelTypes::MEDIUMINT,
         'int' => PropelTypes::INTEGER,
         'integer' => PropelTypes::INTEGER,
         'bigint' => PropelTypes::BIGINT,
